Host-Parasite interaction data from the Global Mammal Parasite Database 2.0

The GMPD 2.0 is a database of the parasites of wild ungulates (artiodactyls and perissodactyls), carnivores, and primates. It comprises 24,000 entries representing data from over 2700 literature sources. Host Latin binomials (column "HostCorrectedName") match those used by the Fritz et al. (2009) mammal supertree (data(mammal_supertree)).

data(gmpd)

Format

An object of class "data.frame".

Source

Ecology Data Papers

References

Stephens, P.R., Pappalardo, P., Huang, S., Byers, J.E., Farrell, M.J., Gehman, A., Ghai, R.R., Haas, S.E., Han, B., Park, A.W., Schmidt, J.P., Altizer, S., Ezenwa, V.O. and Nunn, C.L. (2017), Global Mammal Parasite Database version 2.0. Ecology, 98: 1476-1476. doi:10.1002/ecy.1799

Examples

data(gmpd)